refactor: Separated cosmology file to a submodule directory (#157)
* refactor: Moved cosmology module to separate directory

Moved the 'cosmology' module that is original one file
to its own directory.

* refactor: Moved specific default values to specific modules

* fix: Fix bug on default values

* chore(deps): Pin numba to be <0.59.0
