+++ requirements.txt
@@ -2,6 +2,7 @@ astropy>=5.2.1,<6.0.0
 graphviz==0.20.1
 h5py>=3.8.0
 levmarq_torch==0.0.1
+numba>=0.58.1,<0.59.0
 numpy>=1.23.5
 safetensors>=0.4.1
 scipy>=1.8.0
